qsieve poly
